replace old calendar with new one
